+++ halogen-chromecast/src/iosMain/kotlin/halogen/chromecast/Clock.ios.kt
@@ -1,0 +1,7 @@
+package halogen.chromecast
+
+import platform.Foundation.NSDate
+import platform.Foundation.timeIntervalSince1970
+
+internal actual fun halogenCastNow(): Long =
+    (NSDate().timeIntervalSince1970 * 1000).toLong()